requirements.txt
flask==2.3.2
flask-socketio==5.3.6
python-engineio==4.6.1
python-socketio==5.8.0
eventlet==0.33.3
# 以下是可选依赖，用于更稳定地获取公网IP地址
requests>=2.25.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pydub>=0.25.1
python-dotenv>=1.0.0